requirements.txt
fastapi
uvicorn
pillow
numpy
scikit-learn
joblib
insightface
onnxruntime
python-multipart

